Added payment screen
The buttons on payment screen do not redirect to anywhere right now. The
information on the buttons are also hard coded and needs to be changed.
The graph displaying user progress (i.e. how much money they made in the
past 10 days etc.) has to be inserted (could probably hardcode. Layout
needs to be made more pleasing and follow overall design theme.